feat: cleanup shit
