+++ src/state/slices/opportunitiesSlice/resolvers/uniV2/constants.ts
@@ -1,1 +1,1 @@
-export const FOX_TOKEN_CONTRACT_ADDRESS = '0xc770EEfAd204B5180dF6a14Ee197D99d808ee52d'
+export const TRADING_FEE_RATE = 0.003
